instagram_scraper/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
contourpy==1.1.0
coverage==7.2.7
cryptography==41.0.3
cssselect==1.2.0
cycler==0.11.0
decorator==5.1.1
Django==4.2.3
django-appconf==1.0.5
django-compressor==4.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
exceptiongroup==1.1.2
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
joblib==1.3.1
json5==0.9.14
kiwisolver==1.4.4
loguru==0.7.0
lxml==4.9.3
markdown-it-py==3.0.0
matplotlib==3.7.2
mdurl==0.1.2
multidict==6.0.4
nltk==3.8.1
numpy==1.25.1
openai==0.27.8
outcome==1.2.0
packaging==23.1
parsel==1.8.1
Pillow==10.0.0
playwright==1.35.0
pluggy==1.0.0
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==9.0.4
Pygments==2.15.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.3.2
pytest-cov==4.1.0
pytest-watch==4.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
queuelib==1.6.2
rcssmin==1.1.1
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
rest-framework-simplejwt==0.0.2
rich==13.4.2
rjsmin==1.2.1
scrapeops-scrapy==0.5.2
scrapeops-scrapy-proxy-sdk==1.0
scrapfly-sdk==0.8.6
Scrapy==2.9.0
selenium==4.11.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tld==0.13
tldextract==3.4.4
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
Twisted==22.10.0
typing_extensions==4.7.1
urllib3==2.0.3
w3lib==2.1.2
watchdog==3.0.0
webdriver-manager==4.0.0
wsproto==1.2.0
yarl==1.9.2
zope.interface==6.0
